+++ examples/nextjs/pages/index.tsx
@@ -1,22 +1,14 @@
-import React, { FC } from 'react';
+import React from 'react';
 import { GetStaticProps } from 'next';
-import { DocType, defDocType } from '@component-controls/core';
-import { DocPage } from '@component-controls/app';
-import { Layout, store, getIndexPage } from '@component-controls/nextjs-plugin';
+import {
+  DocPageTemplate,
+  store,
+  getIndexPage,
+} from '@component-controls/nextjs-plugin';
 
-interface PageListProps {
-  type: DocType;
-  docId?: string;
-  storyId?: string;
-}
-
-const HomePage: FC<PageListProps> = ({ type = defDocType, docId, storyId }) => {
-  return (
-    <Layout docId={docId} storyId={storyId}>
-      <DocPage type={type} />
-    </Layout>
-  );
-};
+const HomePage: typeof DocPageTemplate = props => (
+  <DocPageTemplate {...props} />
+);
 
 export const getStaticProps: GetStaticProps = async () => {
   const homePage = getIndexPage(store);